api response APOD arrayList

+++ app/src/main/java/com/david/epay/domain/model/Coordinate.kt
@@ -1,3 +1,3 @@
 package com.david.epay.domain.model
 
-data class Coordinate (val longitude:Long, val latitude:Long)
+data class Coordinate (val longitude:Double, val latitude:Double)
